I'm trying to find hidden videos and photos .
Hi, I'm trying to figure out what videos & photos that are hidden by Google assistance. I  am not the voice on the account.  It is someone who was very close to me.  But they are not with us anymore.  I really could use the help of accessing the memories.  Please.